Adds job title to form. Fixes modal overlay with invalid form msg

+++ layouts/page-layout.tsx
@@ -62,8 +62,10 @@ export default function PageLayout({ children }) {
   };
 
   useEffect(() => {
-    localStorage.clear();
-    router.push("/");
+    if (!isRedtailAuth) {
+      localStorage.clear();
+      router.push("/");
+    }
   }, [isRedtailAuth]);
 
   return (
